Pass active device count.
When running a script it would be helpful to know if there are
still any active devices. This gives the user more information
to determine what they want to perform in their script.

Example: I want to turn on a smart switch when my camera or mic
are active. This switch has a usb 'On Air' light plugged into it.
I don't want the light to turn off until all devices are no
longer active. The active count allows me to check without having
to keep track since Oversight already does this.